requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorlog==4.2.1
idna==2.10
lxml==4.5.2
PyMySQL==0.10.0
pyTelegramBotAPI==3.7.2
pytils==0.3
PyYAML==5.3.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tools==0.1.9
urllib3==1.25.10